requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
backports.zoneinfo==0.2.1
bleach==5.0.0
cffi==1.15.0
channels==3.0.4
constantly==15.1.0
cryptography==37.0.0
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-admin-interface==0.19.0
django-admin-volt==0.0.5
django-adminlte3==0.1.6
django-auto-logout==0.5.0
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.8.0
django-jazzmin==2.5.0
django-modeladmin-reorder==0.3.1
django-summernote==0.8.20.0
django-tinymce==3.4.0
et-xmlfile==1.1.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pytz==2022.1
PyYAML==6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
webencodings==0.5.1
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
